ai-service/requirements.txt
fastapi
google-genai
pydantic
python-dotenv
python-multipart
uvicorn
